+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -4336,6 +4336,12 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                 return translate('power_observer_removed_used_in_fields',
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
+            # workflowAdaptations
+            if 'hide_decisions_when_under_writing__po__{0}'.format(removedRowId) in \
+               self.getWorkflowAdaptations():
+                return translate('power_observer_removed_used_in_fields',
+                                 domain='PloneMeeting',
+                                 context=self.REQUEST)
 
             # check if linked Plone group is empty
             plone_group_id = '{0}_{1}'.format(self.getId(), removedRowId)
@@ -4929,7 +4935,9 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             'waiting_advices_given_advices_required_to_validate':
                 ['waiting_advices_given_and_signed_advices_required_to_validate'],
             'hide_decisions_when_under_writing':
-                ['hide_decisions_when_under_writing_check_returned_to_proposing_group'],
+                ['hide_decisions_when_under_writing_check_returned_to_proposing_group'] +
+                ['hide_decisions_when_under_writing__po__{0}'.format(v['row_id'])
+                 for v in self.getPowerObservers()],
         }
         for base_wfa, dependents in dependencies.items():
             if set(values).intersection(dependents) and base_wfa not in values: