+++ CHANGES.rst
@@ -27,6 +27,10 @@ Changelog
 - Fixed item number not saved when using `Disk` icon to change item number
   on meeting view when item number is a subnumber.
   [gbastien]
+- Added per power observer complementary workflow adaptation to base
+  `hide_decisions_when_under_writing` to let the selected power observers
+  have access to the item decision.
+  [gbastien]
 
 4.2.9rc6 (2024-04-10)
 ---------------------